Deploy MjoLniR with scap3
Based roughly on deployments for ORES and Striker. Each machine
the repo is deployed to will build a virtualenv using the
provided wheels, and then install MjoLniR. MjoLniR isn't deployed
as a wheel to allow for some laziness in versioning, only requiring
a submodule bump (and possibly new wheels, if dependencies change).

Python wheels are basically just zip files, so they are deployed to
archiva with maven and a stdlib only python script. Wheels are placed in
the mirrored repository under org.wikimedia.python with artifact set to
the package name and version the remainder of the wheel version. The
upload script refuses to re-deploy assets which could potentially cause
issues, and instead re-downloads a wheel from archiva if the sha1's
don't match.

Change-Id: I420c8942507fc06c516df09e74e17b90ebf7d2c8