Copy spark config into place on deploy
Putting this configuration into a standard place allows calling spark
commands without having to explicitly point out where mjolnir is
installed to and shortens command lines.

Also bumps mjolnir sub module to master which has support for this
location. Before this can be deployed a puppet patch must be shipped
to create the /etc/mjolnir directory and set its ownership to the
deploy-service so the copy works.

As part of the update to mjolnir master pulls in the new
use-external-memory option for training oversized wikis (enwiki)
and bumps the exectory memory on those from 3G to 4G due to jvm
OOM's seen while training.

Change-Id: I7c69481156f543a8258a2e9b2c90f8e15984caaa