Update spark.yaml for latest mjolnir
Primarily this reworks the data paths so data for various steps
of the pipeline are stored in a shared directory for the run.

Additionally kafka brokers are specified. This doesn't really go here
though, we might need to invent a more generic config file for mjolnir
that everything shares containing information about wiring things
together that has to come from puppet.

Change-Id: I6f2cde6a67fd3a96e616539c14703ad4d8084efc