Switch to simplified virtualenv deployment
The main difference here is that before we would create a single
virtualenv and upgrade it through the deployments. That seems error
prone, instead create a brand new virtualenv inside the deploy
directory and install everything necessary on each deploy.

A symlink is maintained between the old location and the new one
for simplicity. Additionally all the install has moved to the
'fetch' phase of scap which is a better fit for installing various
pieces.

One final benefit of this new method is that rollback should now
work as expected, as long as various things refer to the venv
inside the deploy directory.

Change-Id: I0e2db574a5f8dd1653d353fafe01aec6087e7d7d