Update spark.sh deploy script to use scap script
This was using scap 'command' type, but that generically
passes the command into bash requiring us to know where
it was installed to. Instead use the 'script' type which
only runs scripts included in the repo's scap/scripts/
directory.

Additionally update command that copies spark.yaml into place
to appropriately copy it from the new revision path. This
almost certainly won't undo the copy if we have to rollback,
for that we probably need to switch having scap directly
handle the config deployment.

Change-Id: I04ab375873d96fdf33e44c10f8bca783eda69b7c