make_wheels: Don't upgrade unnecessarily
By doing a fresh install make_wheels was pulling in the latest
versions of each dependency that satisfied mjolnir, but we don't
want to just go upgrading versions of things that work arbitrarily.

Install the frozen requirements before installing mjolnir so they
satisfy the requirements needed. To upgrade a library it needs
to either be updated manually or removed from requirements-frozen.txt.

Tried a couple things but couldn't convince it to stop pulling in new
pip/wheel packages. They were explicitly installed for debian jessie
support which might not be necessary anymore. Future work should check
and remove if necessary. For now upgrading them is less of a problem
than upgrading service dependencies.

Change-Id: Icd6b1d646b498335a08b89273b475afcd1ca9b3d