msearch_daemon: Remove cluster selection/load monitor
This daemon was initially designed with the concept that there would
be an idle cluster it could throw queries at. With mediawiki now
serving read requests from secondary datacenters that is no longer the
case and this mseearch daemon has been sitting idle as it could not
find an unloaded cluster to submit queries to.

The dependent patch reduces the number of workers in each datacenter,
and this patch allows the daemon to query clusters that are currently
serving user queries.

This now has the potential to negatively impact latency for end
user queries. It will need to be monitored on initial rollout and
the the exact parallelism settings adjusted to have minimal impact
on live traffic.

Bug: T318575
Change-Id: If59fe7b005d01ae82389b37a34b6eec425f22255