Repoint shebang lines of conda env
The conda env has all the the shebang lines set to absolute paths on
the build instance, those wont work for us here. Run a replacement
over the first line of all the files in the venv bin to replace
shebangs pointing to the build env with our resulting environmnet.

Change-Id: Ia48b12202b8a43df0d0521536374c118fc8b651c